battery is just 8 bits now

+++ fw/src/battery.h
@@ -3,6 +3,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-extern uint16_t battery_voltage;  // in mV
+extern uint8_t battery_voltage;  // from 0 to 10V
 
 void battery_update();
